+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/operations/diagnostics_settings.py
@@ -5,14 +5,20 @@
 
 
 # pylint: disable=unused-argument
-def create_diagnostics_settings(client, target_resource_id, resource_group=None, logs=None, metrics=None,
-                                namespace=None, rule_name=None, tags=None, event_hub_name=None, storage_account=None,
+def create_diagnostics_settings(client, name, resource_uri,
+                                logs=None,
+                                metrics=None,
+                                event_hub=None,
+                                event_hub_rule=None,
+                                storage_account=None,
                                 workspace=None):
     from azure.mgmt.monitor.models.diagnostic_settings_resource import DiagnosticSettingsResource
 
-    # https://github.com/Azure/azure-rest-api-specs/issues/1058
     parameters = DiagnosticSettingsResource(storage_account_id=storage_account,
-                                            event_hub_name=event_hub_name, metrics=metrics, logs=logs,
-                                            workspace_id=workspace)
+                                            workspace_id=workspace,
+                                            event_hub_name=event_hub,
+                                            event_hub_authorization_rule_id=event_hub_rule,
+                                            metrics=metrics,
+                                            logs=logs)
 
-    return client.create_or_update(target_resource_id, parameters)
+    return client.create_or_update(resource_uri=resource_uri, parameters=parameters, name=name)