+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/disk_encryption.py
@@ -68,7 +68,7 @@ def encrypt_vm(cmd, resource_group_name, vm_name,  # pylint: disable=too-many-lo
     # 1. First validate arguments
 
     if not aad_client_cert_thumbprint and not aad_client_secret:
-        raise CLIError('Please provide either --aad-client-id or --aad-client-cert-thumbprint')
+        raise CLIError('Please provide either --aad-client-cert-thumbprint or --aad-client-secret')
 
     if volume_type is None:
         if vm.storage_profile.data_disks: