+++ tools/automation/verify/verify_packages.py
@@ -7,12 +7,12 @@
 
 import os.path
 import subprocess
-import pip
 import imp
 import glob
 import zipfile
 import logging
 import unittest
+from pkg_resources import working_set
 
 import automation.utilities.path as automation_path
 from automation.utilities.const import COMMAND_MODULE_PREFIX
@@ -43,10 +43,7 @@ class PackageVerifyTests(unittest.TestCase):
     def test_azure_cli_module_installation(self):
         expected_modules = set([n for n, _ in automation_path.get_command_modules_paths(include_prefix=True)])
 
-        pip.utils.pkg_resources = imp.reload(pip.utils.pkg_resources)
-        installed_command_modules = [dist.key for dist in
-                                     pip.get_installed_distributions(local_only=True)
-                                     if dist.key.startswith(COMMAND_MODULE_PREFIX)]
+        installed_command_modules = [dist.key for dist in list(working_set) if dist.key.startswith(COMMAND_MODULE_PREFIX)]
 
         logger.info('Installed command modules %s', installed_command_modules)