+++ src/command_modules/azure-cli-feedback/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.8
++++++
+* Minor fixes.
+
 2.0.7
 ++++++
 * Update for CLI core changes.